Major refactor WIP, dispatch tests asynchronously
Basically done, but a lot of WIP code still in here that I want to have
in git should I have missed some glaring errors during refactoring.

Tests are now run asynchronously, so that each TWAMP test is dispatched
regardless of whether or not we have a response yet. The only exception
is rapid mode, which will just send tests as soon as a response is
received, since the output of the rapid mode hints at Cisco-like
behavior.

Now tests are sent out at 'interval', and we have a separate routine
handling response reading. We also now handle duplicate responses.

Next commit will be removal of WIP code.