Fix mode handling
The final bit of the byte used to encode the mode indicates whether the
reflector is willing to accept using the unauthenticated mode. Since the
library only supports the unauthenticated mode, we need to check that
the final bit is set to 1, as opposed to checking that the byte is equal
to 1. Otherwise, since we're already checking for the unspecified
mode (0), we can report an error if the final bit is not set to 1.